cleaned up code.
1. Added comments for functions to define functionality.
2. Removed unnecessary functions.
